+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @slash-editor/core
 
+## 0.3.1
+
+### Patch Changes
+
+- Fix bubble toolbar flickering when opening Turn into or Ask AI dropdowns
+
 ## 0.3.0
 
 ### Minor Changes